+++ src/slogo/model/Commands/TurtleCommands/SetTowards.java
@@ -7,8 +7,8 @@ import slogo.model.TreeNode;
  * unfinished
  */
 public class SetTowards extends TurtleCommand {
-  private int x;
-  private int y;
+  private double x;
+  private double y;
 
   public SetTowards(String name) {
     super(name);
@@ -16,16 +16,16 @@ public class SetTowards extends TurtleCommand {
 
   @Override
   public void doCommand(TreeNode commandNode) {
-    x = (int)Double.parseDouble(getParamList().get(0));
-    y = (int)Double.parseDouble(getParamList().get(1));
+    x = Double.parseDouble(getParamList().get(0));
+    y = Double.parseDouble(getParamList().get(1));
 
     double angle = Math.toDegrees(Math.atan2(x - turtle.getX(), y - turtle.getY()));
 
     if (angle < 0)
       angle += 360;
 
-    turtle.setDegree((int)angle);
-    int returnValue = Math.abs(turtle.getDegree() - (int)angle);
+    turtle.setDegree(angle);
+    double returnValue = Math.abs(turtle.getDegree() - angle);
     commandNode.setData("" + returnValue);
   }
 }
